feat: add git config message

+++ build.sh
@@ -1,4 +1,8 @@
 #/usr/bin/env bash
+if [[ ! -f ~/.gitconfig ]]; then
+    echo "please configure git first"
+    exit 1
+fi
 cp ~/.gitconfig .
 docker build -t docker_nvim:latest \
     --build-arg CACHEBUST=$(date +%s)  .
